guestbook/requirements.txt
Flask==1.0.2
google-cloud-firestore
google-cloud-datastore
google-auth
mock